+++ apps/web/src/features/organizations/components/Dashboard/index.tsx
@@ -1,8 +1,7 @@
 import MembersCard from '@/features/organizations/components/Dashboard/MembersCard'
 import NewFeaturesCard from '@/features/organizations/components/Dashboard/NewFeaturesCard'
 import OrgsCTACard from '@/features/organizations/components/Dashboard/OrgsCTACard'
-import { Box, Grid2, Stack, Typography } from '@mui/material'
-import SignInButton from '@/features/organizations/components/SignInButton'
+import { Grid2, Stack, Typography } from '@mui/material'
 import Grid from '@mui/material/Grid2'
 import { useOrgSafes } from '@/features/organizations/hooks/useOrgSafes'
 import SafesList from '@/features/myAccounts/components/SafesList'
@@ -17,25 +16,7 @@ import { Link } from '@mui/material'
 import ChevronRightIcon from '@mui/icons-material/ChevronRight'
 import DashboardMembersList from '@/features/organizations/components/Dashboard/DashboardMembersList'
 import { useOrgMembers } from '@/features/organizations/hooks/useOrgMembers'
-import css from './styles.module.css'
-
-const SignedOutState = () => {
-  return (
-    <Box className={css.content}>
-      <Box textAlign="center" p={3}>
-        <Typography variant="h2" fontWeight={700} mb={1}>
-          Sign in to see your organization
-        </Typography>
-
-        <Typography variant="body2" mb={2} color="primary.light">
-          Description
-        </Typography>
-
-        <SignInButton />
-      </Box>
-    </Box>
-  )
-}
+import SignedOutState from '@/features/organizations/components/SignedOutState'
 
 const ViewAllLink = ({ url }: { url: LinkProps['href'] }) => {
   return (
