+++ apps/web/src/features/organizations/components/Members/index.tsx
@@ -7,14 +7,20 @@ import InvitesList from './InvitesList'
 import SearchIcon from '@/public/images/common/search.svg'
 import { useMembersSearch } from '../../hooks/useMembersSearch'
 import { useOrgMembers } from '../../hooks/useOrgMembers'
+import { useAppSelector } from '@/store'
+import { isAuthenticated } from '@/store/authSlice'
+import SignedOutState from '@/features/organizations/components/SignedOutState'
 
 const OrganizationMembers = () => {
   const [openAddMembersModal, setOpenAddMembersModal] = useState(false)
   const [searchQuery, setSearchQuery] = useState('')
+  const isUserSignedIn = useAppSelector(isAuthenticated)
   const { activeMembers, invitedMembers } = useOrgMembers()
 
   const filteredMembers = useMembersSearch(activeMembers, searchQuery)
   const filteredInvites = useMembersSearch(invitedMembers, searchQuery)
+
+  if (!isUserSignedIn) return <SignedOutState />
 
   return (
     <>
